Split classes in their own file
